refactor(common): extract SyncRateLimiter to common/helper/rate-limiter (pv-cug3.2)
Move SyncRateLimiter (and SYNC_PER_SOURCE_HOURLY_LIMIT) from
src/server/calendar/service/import/sync.ts to a new shared helper
at src/server/common/helper/rate-limiter.ts. sync.ts re-exports the
symbols so existing import sites remain unchanged.

This unblocks the AP follow-backfill worker (pv-cug3.4) reusing the
class for per-hostname rate limiting without crossing domain
boundaries into the calendar import domain.

Pure refactor: public API and behavior unchanged.
