+++ src/server/calendar/service/import/sync.ts
@@ -72,18 +72,23 @@ import {
 import { mapVEvent, type MapperOutput } from '@/server/calendar/service/import/mapper';
 import db from '@/server/common/entity/db';
 import { createLogger } from '@/server/common/helper/logger';
+import {
+  SyncRateLimiter,
+  SYNC_PER_SOURCE_HOURLY_LIMIT,
+} from '@/server/common/helper/rate-limiter';
 
 const logger = createLogger('calendar.import.sync');
+
+// Re-export the limiter and its hourly cap so existing call sites (tests,
+// API wiring) keep importing from the sync orchestrator while the
+// implementation lives in the shared common helper. The AP follow-backfill
+// worker (pv-cug3.4) imports the limiter directly from
+// `@/server/common/helper/rate-limiter`.
+export { SyncRateLimiter, SYNC_PER_SOURCE_HOURLY_LIMIT };
 
 // ---------------------------------------------------------------------------
 // Constants
 // ---------------------------------------------------------------------------
-
-/** Maximum Sync-Now calls allowed per source in a sliding 1-hour window. */
-export const SYNC_PER_SOURCE_HOURLY_LIMIT = 4;
-
-/** Milliseconds per hour — the rate-limit window. */
-const RATE_LIMIT_WINDOW_MS = 60 * 60 * 1000;
 
 /** Retention cap for ImportRun rows per source. */
 export const IMPORT_RUN_RETENTION = 50;
@@ -129,47 +134,6 @@ export interface SyncResult {
 export interface SyncInput {
   account: Account;
   importSourceId: string;
-}
-
-// ---------------------------------------------------------------------------
-// Rate limiter — per-source sliding window, in-memory
-// ---------------------------------------------------------------------------
-
-/**
- * Process-local sliding-window counter keyed by importSourceId. The window is
- * 1 hour; the cap is {@link SYNC_PER_SOURCE_HOURLY_LIMIT}. Entries expire
- * lazily on check.
- *
- * An in-memory cap is sufficient for v1 (single-instance deployment is the
- * common case; distributed operators will add Redis-backed limits in a later
- * bead — see complexity-playbook YAGNI).
- */
-export class SyncRateLimiter {
-  private timestamps: Map<string, number[]> = new Map();
-
-  /**
-   * Records a Sync Now attempt. Returns true if the call is allowed.
-   */
-  tryAcquire(sourceId: string, now: number = Date.now()): boolean {
-    const cutoff = now - RATE_LIMIT_WINDOW_MS;
-    const existing = this.timestamps.get(sourceId) ?? [];
-    // Drop stale entries.
-    const fresh = existing.filter(t => t > cutoff);
-    if (fresh.length >= SYNC_PER_SOURCE_HOURLY_LIMIT) {
-      this.timestamps.set(sourceId, fresh);
-      return false;
-    }
-    fresh.push(now);
-    this.timestamps.set(sourceId, fresh);
-    return true;
-  }
-
-  /**
-   * Clear recorded timestamps. Intended for tests.
-   */
-  reset(): void {
-    this.timestamps.clear();
-  }
 }
 
 // ---------------------------------------------------------------------------
